Move fixtures back
